Fix types and licence headers
