+++ osu.Framework/Graphics/Shaders/GlobalPropertyManager.cs
@@ -1,5 +1,5 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
-// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
 using System;
 using System.Collections.Generic;
@@ -64,7 +64,6 @@ namespace osu.Framework.Graphics.Shaders
 
                 global.UnlinkShaderUniform(uniform);
             }
-
         }
     }
 }
