+++ osu.Framework/Graphics/Shaders/IUniform.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
+
+namespace osu.Framework.Graphics.Shaders
+{
+    public interface IUniform
+    {
+        void Update();
+
+        Shader Owner { get; }
+        int Location { get; }
+    }
+}
